make initialization more concise
